+++ index.html
@@ -7,8 +7,7 @@
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кекстаграм</title>
-  <script src="js/main.js></script>
-  <script src="js/data.js" type="module"></script>
+
 </head>
 
 <body>
@@ -233,6 +232,7 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
-
+  <script src="js/main.js" type="module"></script>
 </body>
+
 </html>
